CIP

Documentation

A quick note about the notation in this document: for the functions below, you will see something that looks like

action :: CTD -> CGen -> Double

The "::" means "has type"; what follows is the type signature of the function, i.e. the types of the arguments it takes and the type of its return values. These are all in order and separated by arrows. In the above example, then, we see that the action function takes two arguments (the first of type CTD and the second of type CGen) and returns a Double value (i.e. a decimal number).

Once you get used to their format, type signatures are very useful for showing you how to use a function. The only thing they don't tell you is which argument does what (which can get confusing if you have two arguments of the same type -- but that's what this documentation is for!) If you ever forget the type signature of a function while working, rather than checking back in the documentation, you can type ":t <functionname>" into GHCi, and it will print out the type signature of the function <functionname>.

Data Structure Definitions

type Vertex = (Int, Int)

A Vertex is meant to represent a lattice point. These will be, for instance, vertices of our convex integral paths as well as the points that we count up in order to compute L for a convex generator. A Vertex is made by typing a 2-tuple of integers: for instance,

let vert = (4,5)

type Edge = (Int, Int)

An Edge represents the displacement vector corresponding to an edge of a convex integral path. Being a 2-tuple of integers, an Edge is syntactically identical to a Vertex; Haskell cannot tell the difference between the two. Their names are purely for readability of function names and of code.

data Label

Labels represent the letters attached to edges of a convex generator. As such, there are two types of Labels: E and H. Note that these are not Strings but data types in their own right; we don't type "E" but instead simply E. Also, due to Haskell's naming conventions, the letters must be capitalized.

Constructors

E

H

data CIP

CIP stands for "convex integral path." The data that a CIP holds is simply a list of Vertices, which correspond to every lattice point along the convex integral path (NOT just the ones between two edges of different slopes). That said, you cannot substitute a list of Vertices for a CIP; they are two different data structures.

The function getVertices takes a CIP and returns the list of vertices that is stored in it: for instance,

>>> getVertices (pathFromVerts [(0,2),(2,2),(2,0)])
[(0,2),(1,2),(2,2),(2,1),(2,0)]

(See below for information about how pathFromVerts works.)

data CGen

A CGen represents a convex generator. It contains a convex integral path and a list of labels. The Labels correspond to the labels on the edges of the convex generator, from left to right.

The functions getPath and getLabels both take a CGen as an argument and return the CIP and the list of Labels, respectively, that are held in the CGen.

data CTD

CTD stands for "convex toric domain." Since convex toric domains are precisely the areas in the first quadrant bounded by a CIP, the only data CTDs hold are the CIPs that make up their boundary. However, for ease of use and internal efficiency, there are two more specific types of CTDs. For any floats (i.e. deecimal numbers) a and b, "Ellipsoid a b" represents the ellipsoid E(a,b), and "Polydisk a b" represents the polydisk P(a,b).

Note: the following syntax is required to make a CTD that is not an Ellipsoid or a Polydisk:

let myctd = CTD { getBoundary = <boundary> }

where <boundary> is a CIP. Moreover, getBoundary is a function that take a CTD and returns the CIP that is its boundary. However, getBoundary only works on generic CTDs, NOT on Ellipsoids or Polydisks.

Constructors

Ellipsoid Double Double

Polydisk Double Double

CTD

Fields

getBoundary :: CIP

type Decomposition = ([CGen], [CGen])

A Decomposition represents all the data that is produced from applying Theorem 1.19. Factorizations are represented as lists of convex generators (where each element of the list is a factor); decompositions, therefore, are 2-tuples of factorizations. The first element in the tuple represents the factorization of lambda' (i.e. the convex generator that we apply Theorem 1.19 to); the second element represents the factorization of lambda.

Creating Convex Integral Paths and Convex Generators

pathFromVerts :: [Vertex] -> CIP

Given a list of vertices, this function will create the convex integral path formed by connecting each of the vertices to its successor in the list with a straight line. Note that the vertices must create a valid convex integral path: that is, the first vertex must be on the postive y-axis, the last vertex must be on the positive x-axis, the x-values of the vertices must be increasing, and convexity must be satisfied.

pathFromEdges :: [Edge] -> CIP

Given a list of edges, this function will create the convex integral path formed by connecting each of these edges in the order they are given. Note that the edges must form a valid convex integral path: that is, they must have non-negative x-values and non-positive y-values, and the slopes must be descending.

makeCGen :: CIP -> [Label] -> CGen

Given a CIP and a list of Labels, this function creates the convex generator formed by labeling the first edge (from the left) of the CIP with the first label in the list, the second edge with the second label, and so on. There must be precisely as many labels as there are edges of the CIP, and recall that horizontal and vertical edges cannot be labelled H.

makeAllEGen :: CIP -> CGen

A convenience function which takes a CIP and creates the convex generator formed by labelling all the edges E.

readCGen :: String -> CGen

This function takes in a String representation of a convex generator and creates a convex generator out of it. The schema for String representations is very similar to Michael Hutchings' notation, but since subscripts and superscripts are infeasible on the command line, we write powers as coefficients and write the edge vectors after the labels. For instance: "3E(1,-1)" represents the convex generator with one edge going from (0,3) to (3,0) labelled "E." Note two big differences from Hutchings' notation: first, that the y-value of edges does retain its negative sign, i.e. we have 3E(1,-1) and NOT 3E(1,1); and secondly, that the label letters are capitalized.

Basic Properties of Convex Generators

getX :: Num a => CGen -> a

Computes x() for a given convex generator (i.e. the x-value of the final vertex of the generator).

getY :: Num a => CGen -> a

Computes y() for a given convex generator (i.e. the y-value of the first vertex of the generator).

getH :: Num a => CGen -> a

Computes h() for a given convex generator (i.e. the number of edges labelled H).

getM :: Num a => CGen -> a

Computes m() for a given convex generator (i.e. the total multiplicity of all edges of the generator, which is equal to the number of lattice points on the path of the generator minus one).

getL :: Num a => CGen -> a

Computes L() for a given convex generator (i.e. the total number of lattice points enclosed by the generator, including points on the generator and on the positive axes).

index :: Num a => CGen -> a

Computes the index I() of the generator: for a generator cg,

index cg = 2*(getL cg - 1) - getH cg

action :: CTD -> CGen -> Double

Computes the symplectic action of a given convex generator with respect to a given convex toric domain. Actions with respect to Ellipsoids and Polydisks are computed using the formulas specific to those convex toric domains (see "Beyond ECH Capacities," Example 1.14); general domains are computed using the general action formula (see "Beyond ECH Capacities," Definition 1.13).

Less Than and Theorem 1.19

Condition Checks

isLessThan :: CTD -> CTD -> CGen -> CGen -> Bool

Given two CTDs, ctd and ctd', and two CGens, cg and cg', isLessThan ctd ctd' cg cg' returns True if cg is less than or equal to cg' with respect to ctd and ctd' (see "Beyond ECH Capacities" Definition 1.17) and False otherwise. Note that the first CTD supplied goes with the first CGen supplied, i.e. for the action condition, isLessThan ctd ctd' cg cg' will check that

action ctd cg <= action ctd' cg'

noSharedEOrbits :: CGen -> CGen -> Bool

Given two CGens, this function returns True if the generators share no elliptical orbits and False otherwise.

bullet2 :: Decomposition -> Bool

Checks if a given Decomposition satisfies the second bullet point of Theorem 1.19: that is, for any pair of indices i and j in the factorizations of the Decomposition, either the factors at index i and j are equal in each of the factorizations, or the factors at index i and j in the second factorization (correspondig to lambda, not lambda') share no elliptical orbit.

bullet3 :: Decomposition -> Bool

Checks if a given Decomposition satisfies the third bullet point of Theorem 1.19: that is, for any subset of indices in the factorizations, if we take the product of all the factors at those indices in the lambda factorization and the product of all the factors at those indices in the lambda' factorization, the two resulting convex generators will have equal indices (I() values).

validDecomposition :: CTD -> CTD -> Decomposition -> Bool

Given two CTDs and a Decomposition, checks if the Decomposition satisfies all the conditions of Theorem 1.19. This is equivalent to running the functions bullet2 and bullet3 on the Decomposition, and then going through each of the factorizations and checking that the factors in the right (lambda) factorizations are less than the corresponding factors in the left (lambda') factorization.

Computating Convex Generators, Factorizations, and Decompositions

generateLessThan :: CTD -> CTD -> CGen -> [CGen]

This function computes a list of all convex generators less than or equal to a given CGen with respect to two given CTDs. If you just want to know if one CGen is less than or equal to another, use isLessThan above.

factorizations :: CGen -> [[CGen]]

Computes all factorizations of a given CGen.

decompositions :: CTD -> CTD -> CGen -> [Decomposition]

Computes all the decompositions of a given CGen which satisfy the first bullet point of Theorem 1.19.

thm119 :: CTD -> CTD -> CGen -> Maybe Decomposition

Given two CTDs and a CGen, this function attempts to apply Theorem 1.19 to the CGen by finding a Decomposition that satisfies the conditions of the Theorem. If no valid Decomposition exists, then the value Nothing is returned; otherwise, the value Just decomp is returned, where decomp is a valid Decomposition.

Note: in order to extract the decomposition from this Just decomp format, use the function maybe: for instance, if ctd1 and ctd2 are CTDs and cg is a CGen:

let result = thm119 ctd1 ctd2 cg
let extracted = maybe ([],[]) id result

The above will set the value of extracted to ([],[]) (i.e. an empty Decomposition) if no valid Decomposition exists or to the value of the Decomposition itself (with no "Just" attached) if a valid Decomposition does exist.

thm119' :: CTD -> CTD -> CGen -> [Decomposition]

This function is very similar to thm119, except that it returns a list of all valid Decompositions. if no valid Decomposition exists, then the empty list, [], will be returned.