Scan

Documentation

This module exists provides the ability to scan the results of applying Hutchings' Theorem 1.19 to many different convex toric domains and many different convex generators. The results of scans can be output in many different formats and saved to a file if desired. The expected workflow is: first, create a Scanner (see 'Creating Scanners' below); then, modify the output settings (see 'Modifying Output' below); and finally, use the scan function.

Currently, the only types of allowed scans are on one variable in the first CTD and one variable in the second. For instance, in the problem of embedding P(a,1) into E(c,c), one could scan the results of Theorem 1.19 for many different a and c values. This module also re-exports the CIP module, so that you don't have to load both of them.

Since it is hard to explain how to use the Scanner creation functions without getting into some details about Haskell, I have included an example usage here. My hope is that adapting this example will allow you to use the other available functions even without understanding all of the Haskell. Of course, if you have any questions, please don't hesitate to email me!

Let's use the P(a,1) into E(c,c) problem for our example. Imagine we want to check whether E(i,-i), for i from 1 to 10, will allow us to extend Michael Hutchings' result on this problem (Theorem 1.4 in "Beyond ECH Capacities"). First, we create a list of all our CGens:

let cgs = map (\i -> makeAllEGen (pathFromEdges [(i,-i)])) [1..10]

(Don't worry if you aren't clear on how this works -- you can simply copy this line as I've written it, and if you want to change the range of the i's to, say, everything from x to y inclusive, replace the [1..10] on the end with [x..y].) Now, we need to create a Scanner. We will scan over all a values from 2.41 to 2.70, inclusive, in increments of 0.01. For each a, we will try every c value from 2+a/2-0.05 up to 2+a/2 in increments of 0.01. The syntax for this is:

let s = rangeOver (\a -> Polydisk a 1) (\c -> Ellipsoid c c) cgs (2.41,2.70,0.01) (\a -> 2+a/2) (-0.05,0,0.01)

Here, we use several so-called "lambda functions". For instance, "(\a -> Polydisk a 1)" represents a function that takes in some value of a and returns the CTD given by Polydisk a 1. Now that we have our Scanner, we can call several functions to modify its output:

let s' = (tabulate . noOffsets . toStd) Scanner

(The dots here are the Haskell syntax for function composition.) Now, we are ready to run our scan:

scan s'

module CIP

data ScanRange t1 t2

data Scanner t1 t2

Creating Scanners

normalOver :: (RealFrac t1, Read t1, Show t1, RealFrac t2, Read t2, Show t2) => (t1 -> CTD) -> (t2 -> CTD) -> [CGen] -> [t1] -> [[t2]] -> Scanner t1 t2

Creates a Scanner for a normal scan. The first two arguments are functions that generate CTDs; the third is a list of the CGens to scan with. The fourth and fifth arguments specify all the values that will be put into the first and second arguments, respectively, to yield CTDs to scan over. For instance, in the P(a,1) into B(c,c) example, the fourth argument is the list of all a values, and the fifth argument is a list containing a list of c values to try for each a value. For a given a value, the scan will try each c value in the corresponding list; as soon as it finds a valid Decomposition, it will move on to the next a value. As soon as some a value has a valid Decomposition for its very first c value, the scan will end. Note that this behavior makes sense in practice: for instance, in the P(a,1) into B(c,c) problem, creating an increasing list of c values for each a value ensures that if (a,c) yields a valid decomposition, then that same Decomposition will be valid for every c value after c in the list.

rangeOver :: (RealFrac t1, Read t1, Show t1, RealFrac t2, Read t2, Show t2) => (t1 -> CTD) -> (t2 -> CTD) -> [CGen] -> (t1, t1, t1) -> (t1 -> t2) -> (t2, t2, t2) -> Scanner t1 t2

Creates a Scanner for a range scan. The only difference between this and a normal scan (see normalOver above) is that, rather than specifying each CTD explicitly, the arguments given are a range of values to try. In the P(a,1) into B(c,c) example, we can make the third argument (2.41,2.70,0.01) to scan every a value from 2.41 to 2.70, inclusive, in increments of 0.01. If we want to try, for each a, every c value from 2+a/2-0.05 to 2+a/2 in increments of 0.01, we make the fourth argument (\a -> 2+a/2) and the fifth argument (-0.05,0,0.01).

quickOver :: (RealFrac t1, Read t1, Show t1, RealFrac t2, Read t2, Show t2) => (t1 -> CTD) -> (t2 -> CTD) -> [CGen] -> (t1, t1, t1) -> (t1 -> t2) -> (t2, t2, t2) -> Scanner t1 t2

Creates a Scanner for a quick scan. This works a lot like the range scan; see rangeOver above for details on how the arguments work. However, the quick scan expects the following property to hold for its scan range: given two a values (i.e. values used to create the first CTD, P(a,1) in the Polydisk into Ball example), a1 and a2, and given two corresponding c values, c1 and c2, with the same offset from the largest c values of a1 and a2, respectively (i.e. in the Polydisk into Ball example, take c1 = 2+a1/2-e and c2 = 2+a2/2-e for some e), if there is no valid decomposition given by Theorem 1.19 for a1 and c1 and also no valid decomposition for a2 and c2, then there is no valid decomposition for every a value between a1 and a2 using the corresponding c values with the same offset as before (i.e. for Polydisk into Ball, every a between a1 and a2 has no valid decomposition for c = 2+a/2-e, where e is the same as above). It can be shown that this property holds for the Polydisk into Ball problem. Using this property, quick scan will only fully scan two (a,c) pairs; for the rest, it will take its previous results and refine them by checking only that the action condition for generators being less than each other holds. This massively cuts down on computation time for computationally difficult scans.

Modifying Output

useOffsets :: Scanner t1 t2 -> Scanner t1 t2

Takes a Scanner and returns the same Scanner, but set to use offsets. That is, the Scanner will output the arguments for the second CTD (in the above example, the c values that go into E(c,c)) as offsets from the maximum c value.

noOffsets :: Scanner t1 t2 -> Scanner t1 t2

Takes a Scanner and returns the same Scanner, but set not to use offsets (see useOffsets above).

tabulate :: Scanner t1 t2 -> Scanner t1 t2

Takes a Scanner and returns the same Scanner, but set to print its data to a table. The size of the table will be calculated based on the range of the Scanner when scan is called, and the column headings will default to "CGen," "a." "c," and "Result." See tabulateWith to change these defaults.

tabulateWith :: (String, String, String, String) -> (Int, Int, Int) -> Scanner t1 t2 -> Scanner t1 t2

Given a Scanner, returns the same Scanner, but set to print its data to a table. The first argument is a 4-tuple of the column headings to type; the second argument is a 3-tuple specifying the sizes of the first 3 columns.

plainText :: Scanner t1 t2 -> Scanner t1 t2

Given a Scanner, returns the same Scanner set to print its data plainly. Each result will be printed on its own line; CGens and parameters of the CTDs will be comma-separated, and the results of Theorem 1.19 will follow a colon.

toStd :: Scanner t1 t2 -> Scanner t1 t2

Sets the given Scanner to print its data to StdOut (i.e. the terminal/command line).

toFile :: String -> Bool -> Scanner t1 t2 -> Scanner t1 t2

Sets the given Scanner to print its data to a file. The first argument is a file name, and the second determines whether the file should be overwritten or appended to.

toFileAndStd :: String -> Bool -> Scanner t1 t2 -> Scanner t1 t2

Sets the given Scanner to print its data to both a file and StdOut (i.e. the terminal/command line). The first argument is a file name, and the second determines whether the file should be overwritten or appended to.

Scanning

scan :: (RealFrac t1, Read t1, Show t1, RealFrac t2, Read t2, Show t2) => Scanner t1 t2 -> IO ()

Given a Scanner, runs the scan given by the range of the Scanner and outputs its results in accordance with the output settings of the Scanner.